+++ jsonargparse/__init__.py
@@ -70,4 +70,4 @@ __all__ += _common.__all__
 __all__ += _deprecated.__all__
 
 
-__version__ = "4.29.0"
+__version__ = "4.30.0"